refactor: plural
