+++ src/frontend/src/lib/components/modals/Modals.svelte
@@ -17,7 +17,7 @@
 	import OrbiterTransferCyclesModal from '$lib/components/modals/OrbiterTransferCyclesModal.svelte';
 	import MissionControlTransferCyclesModal from '$lib/components/modals/MissionControlTransferCyclesModal.svelte';
 	import CanisterEditSettingsModal from '$lib/components/modals/CanisterEditSettingsModal.svelte';
-	import SendTokenModal from '$lib/components/modals/SendTokenModal.svelte';
+	import SendTokensModal from '$lib/components/modals/SendTokensModal.svelte';
 
 	let modal: JunoModal | undefined = undefined;
 
@@ -90,6 +90,6 @@
 	<MissionControlTransferCyclesModal on:junoClose={close} detail={modal.detail} />
 {/if}
 
-{#if modal?.type === 'send_token' && nonNullish(modal.detail)}
-	<SendTokenModal on:junoClose={close} detail={modal.detail} />
+{#if modal?.type === 'send_tokens' && nonNullish(modal.detail)}
+	<SendTokensModal on:junoClose={close} detail={modal.detail} />
 {/if}
